[tin 2fbc554c] The last three commits all connect back to a single thread, 6c46b8ae2a7eb757976ab0fe937c68a9c0b35d104522c52b6429be7baf37e95a.
However, when I view them on the web viewer, I notice that each commit links to the _latest_ version of that thread. I would have expected that each would link to the version of the thread that was stored when the commit was made. Are those earlier versions still available? Is it clear in each commit's metadata which version of the thread they were associated with?

If so, let's update the web thread viewer to show the correct version for each commit (along with a note that the thread had other versions in other commits). Then on the thread page, let's show each of the thread's versions.